feat: add AsyncChip component and refactor PullRequestCard and RepositoryCard imports
